feature(gifs): employee delete his/her gif post
[Fixes #10]
